Fix maximum pages calculation by using a custom ceiling function for positive numbers (instead of using math.ceil() from math library to avoid adding new dependencies)
Previously it gives wrong total page numbers, such as zero, as the code does not behave like a ceiling function.

>> "<nav class=\"pagination\" role=\"navigation\">\n\t<span class=\"page-number\">Page 1 of 1</span>\n</nav>"
The original code uses hard coded html output for page 1 to hide this problem. It's now removed. Adding a custom pagination.hbs in the theme partials folder revealed this problem.