Part 1 of SAML integration
This abstracts the session handler away from the code, and implements
one using SAML. This allows you to use google to authenticate your blog.

Google (or another provider) is a lot better because they support two-factor
and have a single password that you remember instead of a different one for
each website.

The only issue remaining here is that there are user preferences and accounts
that are in the database that are not used by the SAML integration.

Some other cleanup:
 - Functions that require a session are wrapped by a RequireSession function
   call. This removes a lot of code in server/admin.go and centralizes the
   check to make sure you have a session.
 - Corrects the error status when you fetch something without a session
 - Stops overwriting the config.json file if you have an error in it. Now,
   it just refuses to start and tells you why.

To configure SAML, you need three things:
 - Your SAML certificate, in a file, pointed to by SAMLCert in the config
 - Your SAML key, in a file, pointed to by SAMLKey
 - The IDP Metadata URL or file, pointed to by SAMLIDPUrl

Once you've set things up, set up your SAML IDP to allow access to the app
for your domain, or use something like auth0 or Okta to create an entry for
it.